Adding some error handling
Adding error handling to check that the args passed in are actually ints and not strings. If we can't parse them then we assume they are strings.

Will make this a better handler that asks for the scores again nicely.
Adding a test to cover this scenario

Also adding a test for the next case I want to work on - entering scores against one player. However, .net core debugger is not playing nicely so debugging is being annoying. Going to come back to it tomorrow